Move output and status information to component
