+++ src/machinable/collection.py
@@ -1452,25 +1452,11 @@ class ComponentCollection(InterfaceCollection):
         status: String, status field: 'started', 'finished', 'alive'
         """
         try:
-            return self.filter(
-                lambda item: item.execution
-                and getattr(item.execution, "is_" + status)()
-            )
+            return self.filter(lambda item: getattr(item, "is_" + status)())
         except AttributeError as _ex:
             raise ValueError(f"Invalid status field: {status}") from _ex
 
 
 class ExecutionCollection(ElementCollection):
-    def status(self, status="started"):
-        """Filters the collection by a status attribute
-
-        # Arguments
-        status: String, status field: 'started', 'finished', 'alive'
-        """
-        try:
-            return self.filter(lambda item: getattr(item, "is_" + status)())
-        except AttributeError as _ex:
-            raise ValueError(f"Invalid status field: {status}") from _ex
-
     def __str__(self):
         return f"Executions <{len(self.items)}>"
