+++ src/machinable/execution.py
@@ -1,7 +1,6 @@
 from typing import Any, Dict, List, Optional, Union
 
 import copy
-import random
 import sys
 
 if sys.version_info >= (3, 11):
@@ -9,36 +8,18 @@ if sys.version_info >= (3, 11):
 else:
     from typing_extensions import Self
 
-import arrow
-from machinable import errors, schema
+from machinable import schema
 from machinable.collection import ComponentCollection
 from machinable.component import Component
-from machinable.element import (
-    Element,
-    defaultversion,
-    extract,
-    get_dump,
-    get_lineage,
-)
+from machinable.element import extract, get_dump, get_lineage
 from machinable.errors import ExecutionFailed
 from machinable.interface import Interface, has_many, has_one
 from machinable.project import Project
 from machinable.schedule import Schedule
 from machinable.settings import get_settings
 from machinable.storage import Storage
-from machinable.types import (
-    DatetimeType,
-    ElementType,
-    TimestampType,
-    VersionType,
-)
-from machinable.utils import (
-    generate_seed,
-    load_file,
-    save_file,
-    sentinel,
-    update_dict,
-)
+from machinable.types import ElementType, VersionType
+from machinable.utils import generate_seed, sentinel, update_dict
 
 
 class Execution(Interface):
@@ -236,147 +217,6 @@ class Execution(Interface):
     @property
     def nickname(self) -> str:
         return self.__model__.nickname
-
-    def mark_started(
-        self, timestamp: Optional[TimestampType] = None
-    ) -> Optional[DatetimeType]:
-        if self.is_finished():
-            return None
-
-        if timestamp is None:
-            timestamp = arrow.now()
-        if isinstance(timestamp, arrow.Arrow):
-            timestamp = arrow.get(timestamp)
-
-        save_file(
-            self.local_directory("started_at"),
-            str(timestamp) + "\n",
-            # starting event can occur multiple times
-            mode="a",
-        )
-
-        return timestamp
-
-    def update_heartbeat(
-        self,
-        timestamp: Union[float, int, DatetimeType, None] = None,
-        mark_finished=False,
-    ) -> Optional[DatetimeType]:
-        if self.is_finished():
-            return None
-        if timestamp is None:
-            timestamp = arrow.now()
-
-        if isinstance(timestamp, arrow.Arrow):
-            timestamp = arrow.get(timestamp)
-
-        save_file(
-            self.local_directory("heartbeat_at"),
-            str(timestamp),
-            mode="w",
-        )
-        if mark_finished:
-            save_file(
-                self.local_directory("finished_at"),
-                str(timestamp),
-            )
-
-        return timestamp
-
-    def output(self, incremental: bool = False) -> Optional[str]:
-        """Returns the output log"""
-        if not self.is_mounted():
-            return None
-        if incremental:
-            read_length = self._cache.get("output_read_length", 0)
-            if read_length == -1:
-                return ""
-            output = self.load_file("output.log", None)
-            if output is None:
-                return None
-
-            if self.is_finished():
-                self._cache["output_read_length"] = -1
-            else:
-                self._cache["output_read_length"] = len(output)
-            return output[read_length:]
-
-        if "output" in self._cache:
-            return self._cache["output"]
-
-        output = self.load_file("output.log", None)
-
-        if self.is_finished():
-            self._cache["output"] = output
-
-        return output
-
-    def created_at(self) -> Optional[DatetimeType]:
-        if self.timestamp is None:
-            return None
-
-        return arrow.get(self.timestamp)
-
-    def started_at(self) -> Optional[DatetimeType]:
-        """Returns the starting time"""
-        if not self.is_mounted():
-            return None
-        return self._retrieve_status("started")
-
-    def heartbeat_at(self):
-        """Returns the last heartbeat time"""
-        if not self.is_mounted():
-            return None
-        return self._retrieve_status("heartbeat")
-
-    def finished_at(self):
-        """Returns the finishing time"""
-        if not self.is_mounted():
-            return None
-        return self._retrieve_status("finished")
-
-    def _retrieve_status(self, field: str) -> Optional[DatetimeType]:
-        fields = ["started", "heartbeat", "finished"]
-        if field not in fields:
-            raise ValueError(f"Invalid field: {field}. Must be on of {fields}")
-        status = load_file(self.local_directory(f"{field}_at"), default=None)
-        if status is None:
-            return None
-        if field == "started":
-            # can have multiple rows, return latest
-            status = status.strip("\n").split("\n")[-1]
-
-        try:
-            return arrow.get(status)
-        except arrow.ParserError:
-            return None
-
-    def is_finished(self):
-        """True if finishing time has been written"""
-        return bool(self.finished_at())
-
-    def is_started(self):
-        """True if starting time has been written"""
-        return bool(self.started_at())
-
-    def is_active(self):
-        """True if not finished and last heartbeat occurred less than 30 seconds ago"""
-        if not self.heartbeat_at():
-            return False
-
-        return (not self.is_finished()) and (
-            (arrow.now() - self.heartbeat_at()).seconds < 30
-        )
-
-    def is_live(self):
-        """True if active or finished"""
-        return self.is_finished() or self.is_active()
-
-    def is_incomplete(self):
-        """Shorthand for is_started() and not (is_active() or is_finished())"""
-        return self.is_started() and not (
-            self.is_active() or self.is_finished()
-        )
 
     def __iter__(self):
         yield from self.executables
