+++ tests/test_component.py
@@ -42,6 +42,27 @@ def test_component(tmp_storage):
     with pytest.raises(errors.MachinableError):
         component.version(["modify"])
 
+    # output
+    c = Component().commit()
+    assert c.output() is None
+    c.save_file("output.log", "test")
+    assert c.output() == "test"
+
+    assert c.output(incremental=True) == "test"
+    c.save_file("output.log", "testt")
+    assert c.output(incremental=True) == "t"
+    assert c.output(incremental=True) == ""
+    c.save_file("output.log", "testt more")
+    assert c.output(incremental=True) == " more"
+
+    c.mark_started()
+    assert c.is_started()
+    c.update_heartbeat()
+    assert c.is_active()
+    c.update_heartbeat(mark_finished=True)
+    assert c.is_finished()
+    assert not c.is_incomplete()
+
     p.__exit__()
 
 
@@ -50,7 +71,7 @@ def test_component_launch(tmp_storage):
     assert not component.is_mounted()
     component.launch()
     assert component.is_mounted()
-    assert component.execution.is_finished()
+    assert component.is_finished()
 
     # multiples
     component = Component()
@@ -65,8 +86,8 @@ def test_component_launch(tmp_storage):
         assert e1.execution is None
         e2 = Component().launch()
         assert e2.execution is None
-    assert e1.execution.is_finished()
-    assert e2.execution.is_finished()
+    assert e1.is_finished()
+    assert e2.is_finished()
 
     class Example(Component):
         def __call__(self):
@@ -146,11 +167,11 @@ def test_component_export(tmp_storage):
         exec(script)
 
     e = Execution().add(component).commit()
-    assert not component.execution.is_started()
+    assert not component.is_started()
 
     exec(script)
 
-    assert component.execution.is_finished()
+    assert component.is_finished()
     assert component.load_file("test_run.json")["success"]
 
     # inline
@@ -160,7 +181,7 @@ def test_component_export(tmp_storage):
     script_filepath = component.save_file("run.sh", script)
 
     print(commandlib.Command("bash")(script_filepath).output())
-    assert component.execution.is_finished()
+    assert component.is_finished()
     assert component.load_file("test_run.json")["success"]
