+++ tests/test_schedule.py
@@ -21,5 +21,5 @@ def test_schedule(tmp_storage):
         with Supported(schedule=["scheduled"]) as execution:
             component = Component().launch()
             assert component.execution is None
-        assert component.execution.is_finished()
+        assert component.is_finished()
         assert execution.schedule.test()
